requirements.txt
qrcode[pil]
Pillow
hypothesis
zxing-cpp
